requirements.txt
alembic==1.16.4
APScheduler==3.11.0
clarifai==2.6.2
emoji==1.7.0
instapy==0.6.16
instapy-chromedriver==2.46
keyring==25.6.0
psycopg==3.2.7
python-dotenv==1.1.1
selenium==3.141.0
sqlalchemy==2.0.41
SQLAlchemy-Utils==0.41.2
urllib3==1.26.18
